fix last changes

+++ packages/alea-frontend/pages/api/access-control/[aclid]/member/[memberid].ts
@@ -4,18 +4,9 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     if (!checkIfDeleteOrSetError(req, res))
         return;
     const userId = await getUserIdOrSetError(req, res);
-    const { memberid, aclid } = req.query;
-    const { updaterACLId, isOpen } = await executeAndEndSet500OnError<any>(`SELECT isOpen,updaterACLId FROM AccessControlList WHERE id=?`, [aclid], res);
-    const isMember = (await executeAndEndSet500OnError<Array<any>>(`SELECT count(AccessControlList.id) as conter FROM AccessControlList
-        INNER JOIN
-            ACLMembership
-        ON 
-            AccessControlList.id = ACLMembership.parentACLId
-        WHERE
-        ACLMembership.memberUserId =? and AccessControlList.id`, [userId, updaterACLId], res))[0].conter;
-    if (isOpen || isMember > 0) {
-        await executeAndEndSet500OnError(`DELETE FROM ACLMembership WHERE id=?;`, [memberid], res);
-        res.status(200).end();
-    }
-    res.status(403).end();
+    const { memberid } = req.query;
+    //TODO: Check the user have right to delete member or a acl to the this acl, this should be done in redis flatten.
+    //TODO: Only the user can delete themself from list.
+    await executeAndEndSet500OnError(`DELETE FROM ACLMembership WHERE id=? and memberUserId=?;`, [memberid, userId], res);
+    res.status(200).end();
 }
